+++ permissions/groups/permissions.ps1
@@ -25,8 +25,7 @@ function Resolve-MS-Entra-ExoError {
         try {
             if (-not [string]::IsNullOrEmpty($ErrorObject.ErrorDetails.Message)) {
                 $httpErrorObj.ErrorDetails = $ErrorObject.ErrorDetails.Message | ConvertFrom-Json
-            }
-            elseif ($ErrorObject.Exception.GetType().FullName -eq 'System.Net.WebException') {
+            } elseif ($ErrorObject.Exception.GetType().FullName -eq 'System.Net.WebException') {
                 if ($null -ne $ErrorObject.Exception.Response) {
                     $streamReaderResponse = [System.IO.StreamReader]::new($ErrorObject.Exception.Response.GetResponseStream()).ReadToEnd()
                     if (-not [string]::IsNullOrEmpty($streamReaderResponse)) {
@@ -37,18 +36,14 @@ function Resolve-MS-Entra-ExoError {
             $errorDetailsObject = $httpErrorObj.ErrorDetails
             if ($errorDetailsObject.error_description) {
                 $httpErrorObj.FriendlyMessage = $errorDetailsObject.error_description
-            }
-            elseif ($errorDetailsObject.error.message) {
+            } elseif ($errorDetailsObject.error.message) {
                 $httpErrorObj.FriendlyMessage = "$($errorDetailsObject.error.code): $($errorDetailsObject.error.message)"
-            }
-            elseif ($errorDetailsObject.error.details.message) {
+            } elseif ($errorDetailsObject.error.details.message) {
                 $httpErrorObj.FriendlyMessage = "$($errorDetailsObject.error.details.code): $($errorDetailsObject.details.message)"
-            }
-            else {
+            } else {
                 $httpErrorObj.FriendlyMessage = $httpErrorObj.ErrorDetails
             }
-        }
-        catch {
+        } catch {
             $httpErrorObj.FriendlyMessage = $httpErrorObj.ErrorDetails
         }
         Write-Output $httpErrorObj
@@ -125,8 +120,7 @@ function Get-MSEntraAccessToken {
 
         $createEntraAccessTokenResponse = Invoke-RestMethod @createEntraAccessTokenSplatParams
         Write-Output $createEntraAccessTokenResponse.access_token
-    }
-    catch {
+    } catch {
         $PSCmdlet.ThrowTerminatingError($_)
     }
 }
@@ -138,8 +132,7 @@ function Get-MSEntraCertificate {
         $rawCertificate = [system.convert]::FromBase64String($actionContext.Configuration.AppCertificateBase64String)
         $certificate = [System.Security.Cryptography.X509Certificates.X509Certificate2]::new($rawCertificate, $actionContext.Configuration.AppCertificatePassword, [System.Security.Cryptography.X509Certificates.X509KeyStorageFlags]::Exportable)
         Write-Output $certificate
-    }
-    catch {
+    } catch {
         $PSCmdlet.ThrowTerminatingError($_)
     }
 }
@@ -159,18 +152,19 @@ try {
 
     $actionMessage = 'Retrieving permissions'
     Write-Information $actionMessage
-    $baseUri = "https://graph.microsoft.com/v1.0/groups?`$filter=groupTypes/any(c:c+eq+'Unified')&`$select=id,displayName,onPremisesSyncEnabled,groupTypes&`$top=999&`$count=true"
+
+    $baseUriM365Group = "https://graph.microsoft.com/v1.0/groups?`$filter=groupTypes/any(c:c+eq+'Unified')&`$select=id,displayName,onPremisesSyncEnabled,groupTypes&`$top=999&`$count=true"
     $microsoftEntraIDM365Groups = [System.Collections.Generic.List[object]]::new()
 
-    $nextLink = $baseUri
+    $nextLink = $baseUriM365Group
     do {
-        $splatParams = @{
-            Uri         = $nextLink
-            Headers     = $headers
-            Method      = 'GET'
-            Verbose     = $false
+        $getMicrosoftEntraIDM365GroupsSplatParams = @{
+            Uri     = $nextLink
+            Headers = $headers
+            Method  = 'GET'
+            Verbose = $false
         }
-        $response = Invoke-RestMethod @splatParams
+        $response = Invoke-RestMethod @getMicrosoftEntraIDM365GroupsSplatParams
         if ($response.Value -is [array]) {
             $microsoftEntraIDM365Groups.AddRange($response.Value)
         } elseif ($response.Value) {
@@ -180,7 +174,7 @@ try {
     } while (-not [string]::IsNullOrEmpty($nextLink))
 
     foreach ($permission in $microsoftEntraIDM365Groups) {
-        $displayName = "Security Group - $($permission.displayName)"
+        $displayName = "M365 Group - $($permission.displayName)"
         $displayName = $displayName.substring(0, [System.Math]::Min(100, $displayName.Length))
         $outputContext.Permissions.Add(
             @{
@@ -191,6 +185,41 @@ try {
             }
         )
     }
+
+    $baseUriSecurityGroup = "https://graph.microsoft.com/v1.0/groups?`$filter=NOT(groupTypes/any(c:c+eq+'DynamicMembership')) and onPremisesSyncEnabled eq null and mailEnabled eq false and securityEnabled eq true&`$select=id,displayName,onPremisesSyncEnabled,groupTypes&`$top=999&`$count=true"
+    $microsoftEntraIDSecurityGroups = [System.Collections.ArrayList]@()
+
+    $nextLink = $baseUriSecurityGroup
+    do {
+        $getMicrosoftEntraIDSecurityGroupsSplatParams = @{
+            Uri     = $nextLink
+            Headers = $headers
+            Method  = 'GET'
+            Verbose = $false
+        }
+
+        $response = Invoke-RestMethod @getMicrosoftEntraIDSecurityGroupsSplatParams
+        if ($response.Value -is [array]) {
+            $microsoftEntraIDSecurityGroups.AddRange($response.Value)
+        } elseif ($response.Value) {
+            $microsoftEntraIDSecurityGroups.Add($response.Value)
+        }
+
+        foreach ($permission in $microsoftEntraIDSecurityGroups) {
+            $displayName = "Security Group - $($permission.displayName)"
+            $displayName = $displayName.substring(0, [System.Math]::Min(100, $displayName.Length))
+            $outputContext.Permissions.Add(
+                @{
+                    displayName    = $displayName
+                    identification = @{
+                        Reference = $permission.id
+                    }
+                }
+            )
+        }
+
+        $nextLink = $response.'@odata.nextLink'
+    } while (-not[string]::IsNullOrEmpty($nextLink))
 } catch {
     $ex = $PSItem
     if ($($ex.Exception.GetType().FullName -eq 'Microsoft.PowerShell.Commands.HttpResponseException') -or